llama-index/requirements.txt
llama-index
llama-index-llms-groq
python-dotenv

#data
#https://zenodo.org/records/8339691